Updated Project Structure
