+++ src/main/java/com/group5/IMDB/repositories/UserReviewRepository.java
@@ -1,0 +1,7 @@
+package com.group5.IMDB.repositories;
+
+public interface UserReviewRepository {
+
+    //PlaceHolder For Repo Class
+
+}
