+++ src/main/java/com/group5/IMDB/services/UserReviewService.java
@@ -1,0 +1,7 @@
+package com.group5.IMDB.services;
+
+public class UserReviewService {
+
+    //Placeholder for Service
+
+}
