+++ src/main/resources/application-dev.properties
@@ -7,4 +7,4 @@ spring.datasource.username=qrmiqvglydgzds
 spring.datasource.password=c14e44d26636f618b273c629bbcf05b0a7c091dc83864cad438cd0280d53c176
 spring.sql.init.platform=postgresql
 
-spring.sql.init.mode=always
+#spring.sql.init.mode=always
